requirements.txt
json
os
websocket-client